pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager application"
authors = ["Igor_Aukhadeev <rumbks@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
environs = "^9.5.0"
dj-database-url = "^0.5.0"
django-heroku = "^0.3.1"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.1"

[tool.poetry.dev-dependencies]
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
